backend/pyproject.toml
[project]
name = "flashback-back"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "elevenlabs[pyaudio]>=1.55.0",
    "fastapi[standard]>=0.115.12",
    "pydantic-ai-slim[mistral]>=0.0.46",
    "pydantic-settings>=2.8.1",
    "pymupdf4llm>=0.0.19",
    "python-dotenv>=1.1.0",
]
